+++ src/main/kotlin/icu/windea/pls/lang/projectView/CwtConfigFilesViewProjectNode.kt
@@ -27,6 +27,7 @@ class CwtConfigFilesViewProjectNode(
         if (!file.isDirectory) return false
         val fileProviders = CwtConfigGroupFileProvider.EP_NAME.extensionList
         fileProviders.forEach f@{ fileProvider ->
+            if (!fileProvider.isEnabled) return@f
             val rootDirectory = fileProvider.getRootDirectory(project) ?: return@f
             if (file == rootDirectory) return true
         }
@@ -36,10 +37,12 @@ class CwtConfigFilesViewProjectNode(
     override fun contains(file: VirtualFile): Boolean {
         val fileProviders = CwtConfigGroupFileProvider.EP_NAME.extensionList
         fileProviders.forEach f@{ fileProvider ->
+            if (!fileProvider.isEnabled) return@f
             val rootDirectory = fileProvider.getRootDirectory(project) ?: return@f
             val relativePath = VfsUtil.getRelativePath(file, rootDirectory) ?: return@f
-            val gameId = relativePath.substringBefore('/')
-            return ParadoxGameType.canResolve(gameId)
+            val directoryName = relativePath.substringBefore('/')
+            val gameTypeId = fileProvider.getGameTypeIdFromDirectoryName(project, directoryName) ?: return@f
+            return ParadoxGameType.canResolve(gameTypeId)
         }
         return false
     }