+++ src/main/kotlin/icu/windea/pls/lang/projectView/CwtConfigGameElementNode.kt
@@ -26,40 +26,49 @@ class CwtConfigGameElementNode(
 
     private fun canRepresent(file: VirtualFile): Boolean {
         if (!file.isDirectory) return false
-        if (file.name != value.gameType.id) return false
-        val rootDir = file.parent ?: return false
+        val gameType = value.gameType
         val fileProviders = CwtConfigGroupFileProvider.EP_NAME.extensionList
         fileProviders.forEach f@{ fileProvider ->
+            if (!fileProvider.isEnabled) return@f
             val rootDirectory = fileProvider.getRootDirectory(project) ?: return@f
-            if (rootDir == rootDirectory) return true
+            val directoryName = fileProvider.getDirectoryName(project, gameType)
+            val nodeFile = rootDirectory.findChild(directoryName) ?: return@f
+            if (!nodeFile.isDirectory) return@f
+            return nodeFile == file
         }
         return false
     }
 
     override fun contains(file: VirtualFile): Boolean {
+        val gameType = value.gameType
         val fileProviders = CwtConfigGroupFileProvider.EP_NAME.extensionList
         fileProviders.forEach f@{ fileProvider ->
+            if (!fileProvider.isEnabled) return@f
             val rootDirectory = fileProvider.getRootDirectory(project) ?: return@f
-            val relativePath = VfsUtil.getRelativePath(file, rootDirectory) ?: return@f
-            val gameId = relativePath.substringBefore('/')
-            return ParadoxGameType.canResolve(gameId)
+            val directoryName = fileProvider.getDirectoryName(project, gameType)
+            val nodeFile = rootDirectory.findChild(directoryName) ?: return@f
+            if (!nodeFile.isDirectory) return@f
+            return VfsUtil.isAncestor(nodeFile, file, false)
         }
         return false
     }
 
     override fun getChildren(): Collection<AbstractTreeNode<*>> {
         if (value == null) return emptySet()
-        val gameTypeId = value.gameType.id
+        val gameType = value.gameType
         val children = mutableSetOf<AbstractTreeNode<*>>()
         val directoryNames = mutableSetOf<String>()
         val fileProviders = CwtConfigGroupFileProvider.EP_NAME.extensionList
         fileProviders.forEach f@{ fileProvider ->
+            if (!fileProvider.isEnabled) return@f
             val rootDirectory = fileProvider.getRootDirectory(project) ?: return@f
-            val dir = rootDirectory.findChild(gameTypeId) ?: return@f
-            dir.children.forEach { file ->
+            val directoryName = fileProvider.getDirectoryName(project, gameType)
+            val nodeFile = VfsUtil.findRelativeFile(rootDirectory, directoryName) ?: return@f
+            if (!nodeFile.isDirectory) return@f
+            nodeFile.children.forEach { file ->
                 if (file.isDirectory) {
                     if (!directoryNames.add(file.name)) return@f
-                    val element = CwtConfigDirectoryElement(project, file.name, value.gameType)
+                    val element = CwtConfigDirectoryElement(project, file.name, gameType)
                     val elementNode = CwtConfigDirectoryElementNode(project, element, settings)
                     children += elementNode
                 } else {